build(orders): initialize orders service with Kafka integration and database setup
This commit sets up the orders service with Kafka integration and configures the
necessary database schema. The service is designed to handle order placements,
cancellations, and trade executions. It includes the creation of relevant
entities, repositories, and event handling mechanisms to ensure smooth
communication with the Kafka messaging system. This foundational work enables
the service to manage orders effectively and prepares it for future enhancements.

+++ services/market-data-consumer/src/main/java/com/tradestream/market_data_consumer/dto/TradeExecuted.java
@@ -1,5 +1,6 @@
 package com.tradestream.market_data_consumer.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -10,8 +11,8 @@ public record TradeExecuted(
     UUID orderId,
     UUID userId,
     String ticker,
-    Double price,
-    Integer quantity,
+    BigDecimal price,
+    BigDecimal quantity,
     String side,
     @JsonFormat(shape = JsonFormat.Shape.STRING)
     Instant timestamp // ISO-8601 UTC
